requirements.txt
flask==1.1.2
gunicorn==20.0.4
prometheus_client
prometheus_flask_exporter
gitpython==3.1.24
toolz==0.11.2
pyyaml==6.0
pytest==6.2.5
pytest-env==0.6.2
requests==2.26.0
pytest-flask
pytest-cov
pylint-flask
pylint
